Updates needed to let clarity adapter use built-in reader
Added a bunch of dev logging to the client that we can turn on and off

Added meta to locations and measurements . This is because sometimes we need to fetch data that is used in building the rest of the data. For example, clarity requires we use a totally different resource just to get the station name.

Added options to the readers. The goal is to be able to pass headers to the fetch method and not have right a whole new method.

Added  a way to pass data from one parser to another in case we need to use that data in the parser somehow.